Restructuring the User and Auction class to allow the User to display their history pruchases and bids

+++ src/org/com1027/coursework/q2/Purchase.java
@@ -15,7 +15,7 @@ public class Purchase {
       return this.buyer;
     }
     
-    public double getQuantityPurchased() {
+    public int getQuantityPurchased() {
       return this.quantityPurchased;
     }
